+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Cleaned up some warnings and unnecessary logging
+- Fixed a bug with how different browsers handle 'indexeddb://' plugin urls - #2
 
 ## [0.6.0] - 2022-08-30